refactor: evolve unsafe casts to idiomatic Rust (Wave 136b)
depot.rs: replace #[allow(cast_sign_loss)] with u64::try_from() —
negative TOML sizes now produce a typed error instead of silent
sign loss. cas_push.rs: evolve PushResult.errors from u64 to usize —
eliminates cast_possible_truncation in commands.rs since error
counts are naturally usize. Production #[allow] count: 13 → 11.

+++ crates/spore-validate/src/cas_push.rs
@@ -149,7 +149,7 @@ impl From<crate::cas::CasEntry> for StoredEntry {
 pub struct PushResult {
     pub stored: u64,
     pub deduplicated: u64,
-    pub errors: u64,
+    pub errors: usize,
     pub error_messages: Vec<String>,
     pub total_bytes_transferred: u64,
     pub elapsed_ms: u64,
@@ -275,7 +275,7 @@ pub fn push_manifest(
     let t0 = Instant::now();
     let mut stored: u64 = 0;
     let mut deduplicated: u64 = 0;
-    let mut errors: u64 = 0;
+    let mut errors: usize = 0;
     let mut error_messages = Vec::new();
     let mut bytes_transferred: u64 = 0;
